+++ R/df_label.R
@@ -3,8 +3,8 @@
 #' Assigns dataset label from a dataset level metadata to a given data frame.
 #' This is stored in the 'label' attribute of the dataframe.
 #'
-#' @param metadata A data frame containing dataset. See 'Metadata' section for
-#'   details.
+#' @param metadata A metacore object or a data frame containing dataset level metadata. See 'Metadata'
+#'   section for details.
 #' @inheritParams xportr_length
 #'
 #' @return Data frame with label attributes.
@@ -41,16 +41,7 @@
 #' adsl <- xportr_df_label(adsl, metadata, domain = "adsl")
 xportr_df_label <- function(.df,
                             metadata = NULL,
-                            domain = NULL,
-                            metacore = deprecated()) {
-  if (!missing(metacore)) {
-    lifecycle::deprecate_stop(
-      when = "0.3.1.9005",
-      what = "xportr_df_label(metacore = )",
-      with = "xportr_df_label(metadata = )"
-    )
-  }
-
+                            domain = NULL) {
   ## Common section to detect default arguments
 
   domain <- domain %||% attr(.df, "_xportr.df_arg_")
@@ -61,6 +52,7 @@ xportr_df_label <- function(.df,
   ## End of common section
 
   assert_data_frame(.df)
+  .df <- group_data_check(.df, verbose = NULL)
   assert_string(domain, null.ok = TRUE)
   assert_metadata(metadata)
 
@@ -79,7 +71,7 @@ xportr_df_label <- function(.df,
     abort("Length of dataset label must be 40 characters or less.")
   }
 
-  if (stringr::str_detect(label, "[^[:ascii:]]")) {
+  if (str_detect(label, "[^[:ascii:]]")) {
     abort("`label` cannot contain any non-ASCII, symbol or special characters.")
   }