+++ R/options.R
@@ -39,8 +39,10 @@
 #'  The default argument for the 'verbose' argument for `xportr_label`.}
 #' \item{xportr.length_verbose}{defaults to `"none"`\cr
 #'  The default argument for the 'verbose' argument for `xportr_length`.}
-#' \item{xportr.type_verbose}{defaults to `"label"`\cr
+#' \item{xportr.type_verbose}{defaults to `"none"`\cr
 #'  The default argument for the 'verbose' argument for `xportr_type`.}
+#' \item{xportr.order_verbose}{defaults to `"none"`\cr
+#'  The default argument for the 'verbose' argument for `xportr_order`.}
 #' \item{xportr.character_types}{defaults to `"character"`\cr
 #'  The default character vector used to explicitly coerce R classes to character XPT types.}
 #' \item{xportr.character_metadata_types}{defaults to `c("character", "char", "text", "date", "posixct", "posixt",
@@ -68,7 +70,7 @@
 #' xportr_options()
 #' @export
 xportr_options <- function(...) {
-  checkmate::assert_subset(names(list(...)), names(xportr_options_list))
+  assert_subset(names(list(...)), names(xportr_options_list))
   if (is.null(names(list(...)))) {
     if (length(list(...)) == 0) {
       queried_options <- names(xportr_options_list)
@@ -86,7 +88,7 @@ xportr_options <- function(...) {
     xportr_options <- grep("^xportr\\.", names(options_list), value = TRUE)
     for (opt in xportr_options) {
       option_value <- options_list[[opt]]
-      do.call(options, stats::setNames(list(option_value), opt))
+      do.call(options, setNames(list(option_value), opt))
     }
   }
 }