+++ R/split.R
@@ -26,6 +26,7 @@
 #'
 #'
 #' @export
+#'
 #' @examples
 #'
 #' adlb <- data.frame(
@@ -35,16 +36,16 @@
 #'
 #' adlb <- xportr_split(adlb, "LBCAT")
 xportr_split <- function(.df, split_by = NULL) {
-  lifecycle::deprecate_warn(
+  deprecate_warn(
     when = "0.4.1",
     what = "xportr_split()",
     with = "xportr_write()",
     details = "Please use the argument `max_gb_size` in the
     function xportr_write() instead` instead."
   )
-
+  .df <- group_data_check(.df)
   attr(.df, "_xportr.split_by_") <- split_by
-  return(.df)
+  .df
 }
 
 #' @rdname xportr_split-deprecated