+++ R/write.R
@@ -9,8 +9,9 @@
 #'   used as `xpt` name.
 #' @param max_size_gb Maximum size in GB of the exported file(s). If size of xpt file exceeds the specified maximum,
 #' it will split the data frame into multiple exported chunk(s).
-#' @param label `r lifecycle::badge("deprecated")` Previously used to to set the Dataset label.
-#' Use the `metadata` argument to set the dataset label.
+#' @param metadata A metacore object or a data frame containing dataset level metadata. See 'Metadata'
+#'   section for details. If provided, `xportr_df_label()` will be called to set
+#'   the dataset label before writing the XPT file.
 #' @param strict_checks If TRUE, xpt validation will report errors and not write
 #'   out the dataset. If FALSE, xpt validation will report warnings and continue
 #'   with writing out the dataset. Defaults to FALSE
@@ -52,18 +53,7 @@ xportr_write <- function(.df,
                          max_size_gb = NULL,
                          metadata = NULL,
                          domain = NULL,
-                         strict_checks = FALSE,
-                         label = deprecated()) {
-  if (!missing(label)) {
-    lifecycle::deprecate_warn(
-      when = "0.3.2",
-      what = "xportr_write(label = )",
-      with = "xportr_write(metadata = )"
-    )
-    assert_string(label, null.ok = TRUE, max.chars = 40)
-    metadata <- data.frame(dataset = domain, label = label)
-  }
-
+                         strict_checks = FALSE) {
   ## Common section to detect default arguments
 
   domain <- domain %||% attr(.df, "_xportr.df_arg_")
@@ -73,8 +63,9 @@ xportr_write <- function(.df,
   # by the user.
 
   ## End of common section
-
+  .df <- group_data_check(.df)
   assert_data_frame(.df)
+
   assert_string(path)
   assert_numeric(max_size_gb, null.ok = TRUE)
   assert_metadata(metadata, null.ok = TRUE)
@@ -82,7 +73,7 @@ xportr_write <- function(.df,
 
   path <- normalizePath(path, mustWork = FALSE)
 
-  name <- tools::file_path_sans_ext(basename(path))
+  name <- file_path_sans_ext(basename(path))
 
   if (!is.null(metadata)) {
     .df <- xportr_df_label(.df, metadata = metadata, domain = domain)
@@ -98,7 +89,7 @@ xportr_write <- function(.df,
 
   checks <- xpt_validate(.df)
 
-  if (stringr::str_detect(name, "[^a-zA-Z0-9]")) {
+  if (str_detect(name, "[^a-zA-Z0-9]")) {
     checks <- c(checks, "`.df` cannot contain any non-ASCII, symbol or underscore characters.")
   }
 
@@ -161,10 +152,10 @@ xportr_write <- function(.df,
 #' @noRd
 get_split_path <- function(path, ind) {
   paste0(
-    tools::file_path_sans_ext(path),
+    file_path_sans_ext(path),
     ind,
     ".",
-    tools::file_ext(path)
+    file_ext(path)
   )
 }
 
@@ -179,7 +170,7 @@ get_split_path <- function(path, ind) {
 export_to_xpt <- function(.df, path, max_size_gb, file_prefix) {
   # Convert GB to bytes
   max_size_bytes <- max_size_gb * 1000^3
-
+  .df <- group_data_check(.df)
   temp_file <- tempfile()
   write_xpt(.df, temp_file)